backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
Pillow>=10.0
gunicorn>=21.0
whitenoise>=6.6
